full run works

+++ slopes/utils.py
@@ -5,17 +5,20 @@ import rioxarray as rxr
 import geopandas as gpd
 import whitebox
 
+
 def translate_to_wbt(pour_points: gpd.GeoSeries, offset: tuple) -> gpd.GeoSeries:
     """
     Translates a points coordinates from the top left of a cell to the center
-    of a cell. 
+    of a cell.
     """
-    return pour_points.translate(xoff=offset[0]/2, yoff=offset[1]/2)
+    return pour_points.translate(xoff=offset[0] / 2, yoff=offset[1] / 2)
 
-def make_dir(path):
-    if os.path.exists(path) and os.path.isdir(path):
+
+def make_dir(path, remove_existing=True):
+    if os.path.exists(path) and os.path.isdir(path) and remove_existing:
         shutil.rmtree(path)
     os.makedirs(path)
+
 
 def setup_wbt(working_dir, verbose=False):
     wbt = whitebox.WhiteboxTools()
